smart_librarian/requirements.txt
chromadb
openai
python-dotenv
streamlit
streamlit-webrtc
av
numpy
pydub
soundfile

